+++ Pandora/Paradigm/Primary/Functor.hs
@@ -1,6 +1,6 @@
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
-module Pandora.Paradigm.Primary.Functor (module Exports, note, this, that, here, there, branches, match) where
+module Pandora.Paradigm.Primary.Functor (module Exports, note, here, branches, match) where
 
 import Pandora.Paradigm.Primary.Functor.Fix as Exports
 import Pandora.Paradigm.Primary.Functor.Equivalence as Exports
@@ -56,10 +56,11 @@ instance Morphable (Into (Right Maybe)) Wye where
 	morphing (extract . run -> Right rs) = Just rs
 	morphing (extract . run -> End) = Nothing
 
-this :: These e ~> Maybe
-this (This x) = Just x
-this (That _) = Nothing
-this (These _ x) = Just x
+instance Morphable (Into (This Maybe)) (These e) where
+	type Morphing (Into (This Maybe)) (These e) = Maybe
+	morphing (extract . run -> This x) = Just x
+	morphing (extract . run -> That _) = Nothing
+	morphing (extract . run -> These _ x) = Just x
 
 that :: These e a -> Maybe e
 that (This _) = Nothing
@@ -71,10 +72,11 @@ here Nowhere = Nothing
 here (Here x) = Just x
 here (There _) = Nothing
 
-there :: Wedge e ~> Maybe
-there Nowhere = Nothing
-there (Here _) = Nothing
-there (There x) = Just x
+instance Morphable (Into (There Maybe)) (Wedge e) where
+	type Morphing (Into (There Maybe)) (Wedge e) = Maybe
+	morphing (extract . run -> Nowhere) = Nothing
+	morphing (extract . run -> Here _) = Nothing
+	morphing (extract . run -> There x) = Just x
 
 branches :: Maybe a -> Maybe a -> Wye a
 branches (Just x) (Just y) = Both x y